requirements.txt
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
pydantic>=2.6
requests>=2.31
tqdm>=4.66
yfinance>=0.2
scikit-learn>=1.4
matplotlib>=3.8
fastapi>=0.111
uvicorn>=0.30

httpx>=0.27
sec-api>=1.0.32
